+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fixed `zephir_preg_match` to use `ZVAL_NULL` instead of `ZEPHIR_NULL`
   [#1946](https://github.com/phalcon/zephir/issues/1946)
+- Fixed `Extension\InternalClassesTest` test to be able run full test suite
+  without Phalcon [#1949](https://github.com/phalcon/zephir/issues/1949)
 
 ## [0.12.7] - 2019-10-03
 ### Fixed